requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.14.2
pydantic==2.6.4
pydantic_core==2.16.3
pydotenv==0.0.7
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-file==2.0.0
setuptools==69.2.0
sniffio==1.3.1
tldextract==5.1.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.2
